arm64: dts: freescale: imx93-phyboard-nash: Set ADC calibration samples
Set number of calibration samples for the i.MX93 ADC. On phyBOARD-Nash,
the ADC_IN1 line is connected to a shunt resistor measuring board power
consumption. The ADC read-out value fluctuates quite a lot during the
calibration resulting in a failed calibration procedure. Thus, lower the
value of calibration samples to pass the calibration process.

Signed-off-by: Primoz Fiser <primoz.fiser@norik.com>