arm64: dts: freescale: imx93-phycore-som: Add ethernet0 alias
Add alias for ethernet0 interface to the phyCORE-i.MX93 SoM include file
The reason behind it is that the physical location of the PHY chip is on
the SoM itself and alias thus belongs in the SoM device-tree.

Moreover, the common bootloader for all phyCORE-i.MX93 based boards
depends on the primary Ethernet interface (ethernet0) to be set to FEC
interface since it is shared among all boards and thus used for Ethernet
support in U-Boot (EQOS is not supported by default in U-Boot due to the
fact that it depends on which eth PHY is used on the carrier board).

Signed-off-by: Primoz Fiser <primoz.fiser@norik.com>